refactor date utils
