+++ condition-checker/src/main/java/com/mersey/rowing/club/condition_checker/view/ThymeleafController.java
@@ -1,0 +1,19 @@
+package com.mersey.rowing.club.condition_checker.view;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class ThymeleafController {
+
+    @GetMapping("/")
+    public String index(Model model) {
+
+        model.addAttribute("currentDateTime", WebsiteDateUtils.getDateNowFormatted());
+        model.addAttribute("previousDays", WebsiteDateUtils.getPrevious7Days());
+        model.addAttribute("nextDays", WebsiteDateUtils.getNext4Days());
+
+        return "index";
+    }
+}
